+++ app/javascript/src/components/Common/Navbar/index.jsx
@@ -1,6 +1,5 @@
 import React from "react";
 
-import { Home, NeetoInsights, UserCircle } from "@bigbinary/neeto-icons";
 import { Sidebar } from "@bigbinary/neetoui/v2/layouts";
 import { Toastr } from "neetoui";
 import { withRouter } from "react-router-dom";
@@ -9,29 +8,7 @@ import authenticationApi from "apis/authentication";
 import { resetAuthTokens } from "apis/axios";
 import { useAuthDispatch } from "contexts/auth";
 
-const navLinks = [
-  {
-    icon: function func() {
-      return <Home size={24} />;
-    },
-    label: "Notes",
-    to: "/notes"
-  },
-  {
-    icon: function func() {
-      return <UserCircle size={24} />;
-    },
-    label: "Contact",
-    to: "/notes"
-  },
-  {
-    icon: function func() {
-      return <NeetoInsights size={24} />;
-    },
-    label: "Settings",
-    to: "/form-elements"
-  }
-];
+import { NAV_LINKS } from "./constants";
 
 const NavBar = () => {
   const authDispatch = useAuthDispatch();
@@ -49,7 +26,7 @@ const NavBar = () => {
   return (
     <Sidebar
       isCollapsed
-      navLinks={navLinks}
+      navLinks={NAV_LINKS}
       profileInfo={{
         dropdownProps: [
           {
