+++ app/javascript/src/components/Dashboard/Notes/Card/CardBody.jsx
@@ -1,0 +1,19 @@
+import React from "react";
+
+import { Typography } from "@bigbinary/neetoui/v2";
+
+const CardBody = () => {
+  return (
+    <>
+      <div className="pb-2">
+        <Typography style="body3">
+          "Are you getting my texts???" she texted to him. He glanced at it and
+          chuckled under his breath. Of course he was getting them, but if he
+          wasn't getting{" "}
+        </Typography>
+      </div>
+    </>
+  );
+};
+
+export default CardBody;
