+++ app/javascript/src/components/Dashboard/Notes/Card/CardFooter.jsx
@@ -1,0 +1,31 @@
+import React from "react";
+
+import { Clock } from "@bigbinary/neeto-icons";
+import { Avatar, Tag, Tooltip, Typography } from "@bigbinary/neetoui/v2";
+
+const CardFooter = ({ imageUrl }) => {
+  return (
+    <>
+      <hr></hr>
+      <div className="flex justify-between pt-4">
+        <Tag label="Getting Started" size="small" color="grey" />
+
+        <div className="flex gap-x-2">
+          <Clock className="my-auto" size={10} />
+          <Tooltip
+            content={<span>Wednesday, 10:30AM</span>}
+            placement="bottom-start"
+          >
+            <Typography style="body3" className="my-auto text-gray">
+              Created 2 hours ago
+            </Typography>
+          </Tooltip>
+          <Avatar user={{ name: "User", imageUrl }} size="medium" />
+        </div>
+        {/* <Clock size={20} color="gray" /> */}
+      </div>
+    </>
+  );
+};
+
+export default CardFooter;
