+++ app/javascript/src/components/Dashboard/Notes/Card/CardHeader.jsx
@@ -1,0 +1,24 @@
+import React from "react";
+
+import { MenuVertical } from "@bigbinary/neeto-icons";
+import { Dropdown, Typography } from "@bigbinary/neetoui/v2";
+
+const CardHeader = () => {
+  return (
+    <>
+      <div className="flex justify-between">
+        <Typography style="h5">How to claim the warranty?</Typography>
+        <Dropdown
+          icon={() => <MenuVertical size={20} color="gray" />}
+          position="bottom-start"
+          buttonStyle="text"
+        >
+          <li>Edit</li>
+          <li>Delete</li>
+        </Dropdown>
+      </div>
+    </>
+  );
+};
+
+export default CardHeader;
